Add OGN to Tokens.js
Hi! I've added to OGN to the list of verified tokens. One of our users was asking about a warning message he was seeing on [uniswap.exchange](https://uniswap.exchange/swap/0x8207c1ffc5b6804f6024322ccf34f29c3541ae26), and it looks like is the way to take care of it for him. If you have any questions, or if this is the wrong way to submit this, give me a shout!

Token contract:

https://etherscan.io/token/0x8207c1ffc5b6804f6024322ccf34f29c3541ae26

Uniswap contract (with working transactions):

https://etherscan.io/address/0x48dc4d621c1b9f7ad919daab69100b3a8834458f